added asm wrapper because it works better

+++ student-distrib/keyboard.c
@@ -22,17 +22,16 @@ char scan_code_1[256] = {
     };
 
 //initalize keyboard
-void keyboard_init(){
+void keyboard_init() {
     enable_irq(KB_IRQ);
 }
 
-
-void keyboard_handler(){
+void keyboard_handler() {
     cli();
     char byte = inb(KB_PORT);
     //only look for keypresses if value is defined (0x39 is last defined value in table above)
     if (byte <= 0x39){
-        char character = scan_code_1[byte];
+        char character = scan_code_1[(int)byte];
         putc(character);
     }
     send_eoi(KB_IRQ);
